# follow-up on #13150: improve the closure-computation algorithm again