--- a/Changes
+++ b/Changes
@@ -241,8 +241,8 @@ _______________
 - #13103: FreeBSD/amd64: properly annotate .o files with non-executable stack
   notes (Konstantin Belousov, review by Nicolás Ojeda Bär)
 
-- #13150: improve a transitive-closure computation algorithm in the flambda
-  middle-end to avoid a compilation time blowup on Menhir-generated code
+- #13150, #????: improve a flambda transitive-closure computation algorithm
+  to avoid a compilation time blowup on Menhir-generated code
   (Florian Weimer, review by Gabriel Scherer and Pierre Chambart,
    report by Richard Jones)