@gasche 2024-05-09
One may reasonably wonder whether it is worth having an algorithm that is faster in the "hard" cases, if it has worse constant factors in the "easy" case (2x slowdown). Let me explain why I think it is worth it.

1. The instances in human-written code are both "easy" and *very small*, so the cost of this closure computation is neglectible compared to the rest of the compilation, and the constant factor does not matter.

2. The issue with this function is thus not its computation time in the easy case, but the lack of robustness that arise from bad asymptotic behavior in harder cases. Performing asymptotically well is more important than being fast.

3. The current algorithm in trunk is inefficient in situations that very plausibly occur in generated code: think of N mutually-recursive functions that all share an invariant first parameter: `let rec f env ... = ... and g env ... = ... and ...`. In this case the current algorithm will perform N times the work of the algorithm in this PR.