@gasche 2024-05-15
@fweimer-rh I wonder if you would be willing to give a look at the code and give your impression on the complexity/performance tradeoff.