@lthls 2024-05-19
I had some discussions with @chambart earlier this week about this issue (the original #13150, not this new algorithm), and we suspect that for menhir-generated parsers we should simply not compute the transitive closure at all (we expect that very few parameters will be invariant or unused, and for the invariant parameters it's unlikely we will have any opportunity to actually make use of that).
However, we didn't find a good way to detect this case so we didn't make any concrete proposals.
Some solutions we thought about:
- a dedicated flag passed to the compiler to warn about big recursive nests (impractical because `dune` doesn't make it easy to pass per-file flags, although since it knows about menhir it might be able to do it automatically for generated parsers)
- a cut-off in the compiler that only computes transitive closures for "small" nests (say, less than 20 functions)

I don't have any code for these suggestions at the moment, but if someone is interested in testing them I could put it on my todo list.