@lthls 2024-05-19
The per-file flags could be made to work independently of the build system with the introduction of unit headers (https://github.com/ocaml/RFCs/pull/26). For example menhir would add at the start of the generated file an attribute disabling this optimisation, and the compiler would read this attribute and change its settings accordingly.
Otherwise I agree that it's not realistic (as far as I know it isn't even possible in dune, you would have to move the files to independent libraries).