@gasche 2024-05-19
(I tried using hashtables for the `state` map and `visited` set, and the performance gains are very small on the synthetic benchmarks -- 30% faster for "isolated" relations, and the same speed for all others. It is not worth the small extra complexity of using different data structures to represent the same data.)