@gasche 2024-06-05
This PR is sitting in a suspended state due to a lack of reviewers, despite demonstrating an improvement over the current trunk on realistic use-cases. Any volunteers to review the implementation for correctness? (This is standard algorithmic stuff, the extra function is a graph traversal that remembers which nodes it already visited to avoid/detect cycles.) (I'm sure @lthls knows plenty of people who could do this for breakfast.)