@lthls 2024-06-05
In my opinion this is very low priority. It's only a marginal improvement for a very marginal use-case that we hope will disappear in the not-too-distant future. I'll still mention it to people around me, some of them love theoretical questions that have little practical impact.