@gasche 2024-08-01
I found out by chance (reading the make build logs) that there is a `utils/strongly_connected_components.ml` module. If no one is willing to review my lovely code, I could maybe try to understand whether this makes the problem at hand much simpler -- intuitively it should be both easy and efficient to compute the transitive closure from a SCC decomposition of the relation.

> In my opinion this is very low priority. It's only a marginal improvement for a very marginal use-case that we hope will disappear in the not-too-distant future. I'll still mention it to people around me, some of them love theoretical questions that have little practical impact.

For the record, I think that you have the wrong perspective. Coccinelle is an important tool in the Linux development community, possibly the only OCaml program that most Linux hackers will ever use. Enabling flambda (the codebase that the "people around [you]" are responsible for) makes the build time for Coccinelle go from "a few seconds" to "34 minutes". This sucks, and it was annoying enough to make Red Hat engineers implement a patch to the OCaml compiler, that went back to "a few seconds" territory. I did the review work myself, and this huge issue mostly went away thanks to their contribution. The PR I propose reduces the bottleneck further, "moving from 5.7 seconds to one second" according to their test. I wouldn't say this is a world-changing patch, but it is nothing to sneer at.

The alternative that you propose is unclear; either wait until flambda is gone for good, or maybe wait for you and Pierre to implement an alternative approach for flambda1, but you're not really sure what would be a good design and no one seems motivated to implement it. That does not sound like a very proposal to me.