@lthls 2024-08-01
I'm proposing to disable this closure computation in cases with too many functions.
@chambart told me that it's very unlikely we will discover anything useful here, particularly on menhir-generated parsers.