@gasche 2024-08-01
Is someone planning to submit a PR that does this?