@lthls 2024-08-01
To clarify a bit: the main goal of this piece of code is to discover invariant parameters. If we have invariant parameters in a set of recursive functions, then it means that occurrences of these functions can be specialised to the invariant parameter. (Typically, `List.map` has the function as an invariant parameter, so if we see it applied to a known function we can specialise it, which allows us to inline the mapped function  if we want to).
For big sets of functions, this is much less interesting because you have to duplicate a lot more code (all of the other functions in the recursive set) when you specialise one function. Moreover, we suspect that huge sets of recursive functions are less likely to feature invariant parameters (we haven't run actual experiments though).
So for these reasons, I believe that approximating the result of the analysis with a sound-but-not-complete approximation (no invariant parameters) when the set of functions grow beyond a certain size (10 functions, maybe ?) is likely a very good compromise.
I'm on holidays until the 12th, at which point I can look at submitting a PR.