@lthls 2024-08-14
> Is someone planning to submit a PR that does this?

Done in #13379.