@lthls 2024-08-30
@gasche I'm considering reviewing this. I have already noted that I'm not moved by the performance argument, which I think is better solved by my own PR, but I'm fine with accepting this PR on other grounds. Would you like to see this make progress even if performance is irrelevant ?